added cube_match_internal and pulled common c routines to cube_utils.c
